Separated database and discord bot logic, cleaned up edgecases

+++ bot.py
@@ -3,8 +3,9 @@ import os
 from dotenv import load_dotenv
 from discord.ext import commands
 import pyrebase
+import firebase_funcs
 
-# Setup Stuff
+# Discordpy setup
 load_dotenv()
 
 TOKEN = os.getenv('DISCORD_TOKEN')
@@ -16,9 +17,7 @@ help_command = commands.DefaultHelpCommand(
 
 bot = commands.Bot(command_prefix='!', intents=intents, help_command=help_command)
 
-number_emojis = ('1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣', '6️⃣', '7️⃣', '8️⃣', '9️⃣', '🔟')
-yes_no_emojis = ('❌','✅')
-
+# Pyrebase setup
 config = {
     "apiKey": os.getenv("FBASE_APIKEY"),
     "databaseURL": os.getenv("FBASE_DATABASEURL"),
@@ -33,76 +32,22 @@ config = {
 firebase = pyrebase.initialize_app(config)
 db = firebase.database()
 
-
-def is_unique_list(list_name, server_id):
-    if not db.child().get().val():
-        return True
-    elif str(server_id) not in list(db.child().get().val()):
-        # Fix by adding server table on bot entry to server
-        return True
-
-    return list_name not in list(db.child(server_id).get().val().keys())
+# constants
+number_emojis = ('1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣', '6️⃣', '7️⃣', '8️⃣', '9️⃣', '🔟')
+yes_no_emojis = ('❌','✅')
 
 
-def create_list(list_name, author_id, server_id, is_for_all=False):
-    server_object = db.child(server_id).get().val()
-
-    data = {
-        "creator": author_id,
-        "members": [author_id],
-        "message_id": "",
-        "for_all": is_for_all
-    }
-    if server_object is None:
-        db.child(server_id).child(list_name).set(data)
-        return db.child(server_id).child(list_name).get().val()
-
-    db.child(server_id).child(list_name).set(data)
-    return db.child(server_id).child(list_name).get().val()
-
-
-def get_list_by_name(list_name, server_id):
-    ret = db.child(server_id).child(list_name).get()
-    return ret.key(), ret.val()
-
-
-def get_list_by_message_id(message_id, server_id):
-    lists = db.child(server_id).get().val()
-    for key, val in lists.items():
-        if val["message_id"] == message_id:
-            return key, val
-    return None
-
-
-def get_all_list_names_by_user(author_id, server_id):
-    output = []
-
-
-    if str(server_id) not in list(db.child().get().val()):
-        return output
-
-    all_lists = db.child(server_id).get().val()
-    for lst in all_lists:
-        if author_id in all_lists[lst]["members"]:
-            output.append(lst)
-    return output
-
-def update_list(list_name, server_id, the_list):
-    return db.child(server_id).child(list_name).update(the_list)
-
-
-def remove_list(list_name, server_id):
-    return db.child(server_id).child(list_name).remove()
-
-def add_new_user_to_public_lists(user_id, server_id):
-    all_lists = db.child(server_id).get()
-    for lst in all_lists.each():
-        if lst.val()["for_all"] and user_id not in lst.val()["members"]:
-            lst.val()["members"].append(user_id)
-            update_list(lst.key(), server_id, lst.val())
-
-
+# helper functions
 def make_output_list(the_list, list_title):
+    """Creates a numbered list of ToDo items with a title
+
+    Args:
+        the_list(List): The list of items in TodoList
+        list_title(String): Title of the list
+
+    Returns:
+        List: formated List of items
+    """
     if len(the_list) > 0:
         output = [f"__**{list_title}**__"]
         for index, item in enumerate(the_list):
@@ -111,7 +56,15 @@ def make_output_list(the_list, list_title):
     return ["All done!"]
 
 
-def make_completed_list(the_list):
+def make_completed_list(the_list) -> list:
+    """Creates a list of completed items
+
+    Args:
+        the_list(List): The list of items in a completed list
+
+    Returns:
+        list: formated List of items
+    """
     if len(the_list) > 0:
         output = [f"__**Completed**__"]
         for index, item in enumerate(the_list):
@@ -120,33 +73,52 @@ def make_completed_list(the_list):
     return []
 
 async def add_emojis(the_message, the_list):
+    """Adds numbered emojis reactions to a message based on a ToDo list
+
+    Args:
+        the_message(discord.Message): The message to add reactions to
+        the_list(List): The list of items in TodoList
+    """
     for index, items in enumerate(the_list):
         await the_message.add_reaction(number_emojis[index])
 
 
 async def add_yes_no_emojis(the_message):
+    """Adds yes/no emojis reactions to a message
+
+    Args:
+        the_message(discord.Message): The message to add reactions to
+
+    Returns:
+        None
+    """
     await the_message.add_reaction(yes_no_emojis[1])
     await the_message.add_reaction(yes_no_emojis[0])
-
 
 # -----Commands-----
 
 # Create new list
 @bot.command(name="makelist", help="Make a new ToDo list")
 async def make_new_list(ctx: discord.ext.commands.Context, permission, *, list_name=None):
+    """Bot command that creates a new ToDo list and adds it to the database
 
+    Args:
+        ctx(discord.ext.commands.Context): The context of the command from Discord
+        permission(String): The permission of the new list. 'all' = public
+        list_name(String): The name of the new list
+    """
     if list_name is None:
         list_name = permission
-        if not is_unique_list(list_name, ctx.guild.id):
+        if not firebase_funcs.is_unique_list(db, list_name, ctx.guild.id):
             await ctx.send("List name already in use")
             return
-        new_list = create_list(list_name, ctx.author.id, ctx.guild.id)
+        new_list = firebase_funcs.create_list(db, list_name, ctx.author.id, ctx.guild.id)
 
     elif permission == "all":
-        if not is_unique_list(list_name, ctx.guild.id):
+        if not firebase_funcs.is_unique_list(db, list_name, ctx.guild.id):
             await ctx.send("List name already in use")
             return
-        new_list = create_list(list_name, ctx.author.id, ctx.guild.id, True)
+        new_list = firebase_funcs.create_list(db, list_name, ctx.author.id, ctx.guild.id, True)
         
         # Add all members since permission all
         for member in ctx.guild.members:
@@ -160,12 +132,18 @@ async def make_new_list(ctx: discord.ext.commands.Context, permission, *, list_n
             """)
         return
 
-    update_list(list_name, ctx.guild.id, new_list)
+    firebase_funcs.update_list(db, list_name, ctx.guild.id, new_list)
     await ctx.message.add_reaction('✅')
 
 
 @bot.command(name="removelist", help="Removes list by name")
 async def remove_list_by_name(ctx: discord.ext.commands.Context, *, list_name=None):
+    """Bot command that removes a list from the database by name
+
+    Args:
+        ctx(discord.ext.commands.Context): The context of the command from Discord
+        list_name(String): Title of the list that is to be deleted
+    """
     if list_name is None:
         await ctx.send("""
             ⚠Invalid command used to remove list.⚠
@@ -174,12 +152,12 @@ async def remove_list_by_name(ctx: discord.ext.commands.Context, *, list_name=No
         return
 
     user_id = ctx.author.id
-    _list_name, the_list = get_list_by_name(list_name, ctx.guild.id)
+    _list_name, the_list = firebase_funcs.get_list_by_name(db, list_name, ctx.guild.id)
     if the_list is None:
         await ctx.send("⚠No list by that name ⚠")
         return
     if user_id in the_list["members"]:
-        new_message = await ctx.send(f"Are you sure you want to delete this list? ```{list_name}``` ✅=YES ❌=NO")
+        new_message = await ctx.send(f"Are you sure you want to delete this list? ```{list_name}```")
         await add_yes_no_emojis(new_message)
     else:
         await ctx.send("⛔You are not a member of this list.⛔")
@@ -188,7 +166,12 @@ async def remove_list_by_name(ctx: discord.ext.commands.Context, *, list_name=No
 # Show users lists
 @bot.command(name="mylists", help="Display ToDo lists for specific user")
 async def show_my_lists(ctx: discord.ext.commands.Context):
-    user_lists_names = get_all_list_names_by_user(ctx.author.id, ctx.guild.id)
+    """Bot command that shows all avaiable list names that are avaiable to the user
+
+    Args:
+        ctx(discord.ext.commands.Context): The context of the command from Discord
+    """
+    user_lists_names = firebase_funcs.get_all_list_names_by_user(db, ctx.author.id, ctx.guild.id)
     if len(user_lists_names) > 0:
         await ctx.send("**Here are your lists:**")
         await ctx.send("\n".join(user_lists_names))
@@ -199,6 +182,12 @@ async def show_my_lists(ctx: discord.ext.commands.Context):
 # Show List Command
 @bot.command(name="showlist", help="Display a ToDo list in chat")
 async def show_list(ctx: discord.ext.commands.Context, *, list_name=None):
+    """Bot command that shows the ToDo list items from a list name
+
+    Args:
+        ctx(discord.ext.commands.Context): The context of the command from Discord
+        list_name(String): The name of the list to be shown
+    """
     if list_name is None:
         await ctx.send("""
             ⚠Invalid command used to show list.⚠
@@ -207,11 +196,11 @@ async def show_list(ctx: discord.ext.commands.Context, *, list_name=None):
         return
 
     user_id = ctx.author.id
-    _list_name, the_list = get_list_by_name(list_name, ctx.guild.id)
+    _list_name, the_list = firebase_funcs.get_list_by_name(db, list_name, ctx.guild.id)
     if the_list is None:
         await ctx.send("⚠No list by that name ⚠")
         return
-    if user_id in the_list["members"]:  # if user_id in get_list_by_name("list_name")["members"]
+    if user_id in the_list["members"]:  # if user_id in firebase_funcs.get_list_by_name(db, "list_name")["members"]
 
         if "todoList" not in the_list:
             await ctx.send("Add some tasks!")
@@ -224,7 +213,7 @@ async def show_list(ctx: discord.ext.commands.Context, *, list_name=None):
         output = todo_output + completed_output
         the_message = await ctx.send("\n".join(output))
         the_list["message_id"] = the_message.id
-        update_list(list_name, ctx.guild.id, the_list)
+        firebase_funcs.update_list(db, list_name, ctx.guild.id, the_list)
         await add_emojis(the_message, the_list["todoList"])
     else:
         await ctx.send("⛔You are not a member of this list.⛔")
@@ -233,25 +222,51 @@ async def show_list(ctx: discord.ext.commands.Context, *, list_name=None):
 # Add item Command
 @bot.command(name="add", help="Add new item to list. Need quotes around list name")
 async def add_item(ctx: discord.ext.commands.Context, list_name, *, new_item):
-    _list_name, the_list = get_list_by_name(list_name, ctx.guild.id)
+    """Bot command that adds a ToDo item to a list
+
+    Args:
+        ctx(discord.ext.commands.Context): The context of the command from Discord
+        list_name(String): The name of the list to be shown
+        new_item(String): The new item that will be added to the list
+    """
+    _list_name, the_list = firebase_funcs.get_list_by_name(db, list_name, ctx.guild.id)
+
+    if _list_name is None and the_list is None:
+        return
+    
     if the_list is None:
         await ctx.send("⚠No list by that name ⚠")
         return
+
     if "todoList" not in the_list:
         the_list["todoList"] = [new_item]
+    elif len(the_list["todoList"]) >= 10:
+        await ctx.send("Max items reached. 10/10 Be more productive.")
+        return
     else:
         the_list["todoList"].append(new_item)
-    update_list(list_name, ctx.guild.id, the_list)
+    firebase_funcs.update_list(db, list_name, ctx.guild.id, the_list)
     await ctx.message.add_reaction('✅')
 
 
 # ----Events-----
 @bot.event
 async def on_member_join(new_member):
-    add_new_user_to_public_lists(new_member.id, new_member.guild.id)
+    """Event that listens for when a new member joins a server, and adds that user to any public listst
+
+    Args:
+        new_member(discord.Member): The new user that joined the guild.
+    """
+    firebase_funcs.add_new_user_to_public_lists(db, new_member.id, new_member.guild.id)
 
 @bot.event
 async def on_command_error(ctx, error):
+    """Event that listens for when a command raises an error
+
+    Args:
+        ctx(discord.ext.commands.Context): The context of the command from Discord
+        error(String): The error that got raised
+    """
     await ctx.send(error)
     await ctx.send("Check available commands below:")
     await ctx.send_help()
@@ -260,11 +275,24 @@ async def on_command_error(ctx, error):
 # Listens for emoji reactions
 @bot.event
 async def on_reaction_add(reaction: discord.Reaction, user: discord.Member):
+    """Event that listens for emoji reactions and responds.
+
+    Args:
+        reaction(discord.Reaction): The reaction object
+        user(discord.Member): The memebr that did that reactions
+    """
     if user.id == bot.user.id:
         return
+    if reaction.emoji not in number_emojis and reaction.emoji not in yes_no_emojis:
+        return
+        
     # Emoji Number reaction 
     if reaction.emoji in number_emojis:
-        list_name, the_list = get_list_by_message_id(reaction.message.id, reaction.message.guild.id)
+        list_name, the_list = firebase_funcs.get_list_by_message_id(db, reaction.message.id, reaction.message.guild.id)
+
+        if list_name is None and the_list is None:
+            return
+
         if user.id not in the_list["members"]:
             await reaction.message.remove_reaction(reaction.emoji, user)
             return
@@ -284,7 +312,7 @@ async def on_reaction_add(reaction: discord.Reaction, user: discord.Member):
         else:
             the_list["completed"].append(item_to_move)
 
-        update_list(list_name, reaction.message.guild.id, the_list)
+        firebase_funcs.update_list(db, list_name, reaction.message.guild.id, the_list)
 
         todo_output = make_output_list(the_list["todoList"], list_name)
         completed_output = make_completed_list(the_list["completed"])
@@ -295,13 +323,24 @@ async def on_reaction_add(reaction: discord.Reaction, user: discord.Member):
     
     # YES NO reaction
     elif reaction.emoji in yes_no_emojis:
+        if "`" not in reaction.message.content:
+            # Handle 'race' condition
+            return
+
         list_name = reaction.message.content.split("```")[1]
-        _list_name, the_list = get_list_by_name(list_name, reaction.message.guild.id)
+        _list_name, the_list = firebase_funcs.get_list_by_name(db, list_name, reaction.message.guild.id)
+
+        if the_list is None:
+            # Handle 'race' condition
+            return
+            
         if user.id not in the_list["members"]:
             await reaction.message.remove_reaction(reaction.emoji, user)
             return
         if reaction.emoji ==  yes_no_emojis[1]:
-            remove_list(list_name, reaction.message.guild.id)
+            firebase_funcs.remove_list(db, list_name, reaction.message.guild.id)
+            await reaction.message.edit(content=f"{list_name} has been deleted!")
+            await reaction.message.clear_reactions()
         elif reaction.emoji == yes_no_emojis[0]:
             await reaction.message.delete()
